refactor: move preserveStrings & restoreStrings
